ci: reviewdogとgolangci-lintによるチェックを追加

+++ internal/model.go
@@ -1,20 +1,20 @@
 package internal
 
 type Mapping struct {
-	Day   uint
-	Index int
+	Day   uint32
+	Index uint8
 }
 
-func ConvertDataset(mappings string) (map[uint]int, []Mapping) {
+func ConvertDataset(mappings string) (map[uint32]uint8, []Mapping) {
 	mappingLen := len(mappings) >> 1
 	results := make([]Mapping, mappingLen)
-	holidays := make(map[uint]int, mappingLen)
-	day := uint(0)
+	holidays := make(map[uint32]uint8, mappingLen)
+	day := uint32(0)
 	j := 0
 	for i := range mappingLen {
-		day += uint(mappings[j])
+		day += uint32(mappings[j])
 		j++
-		index := int(mappings[j])
+		index := mappings[j]
 		j++
 
 		holidays[day] = index
